Updated font inside of about screen. Added button image to back button in about sceen.

+++ Minute Math Racer/Minute Math Racer/PracticeLevelView.m
@@ -418,7 +418,7 @@
 
     //generates random number between 0 and 1
     questionOrientation =arc4random()%2;
-    NSLog(@"Question Orientaiton: %d",questionOrientation);
+    NSLog(@"Question Orientaiton: %ld",(long)questionOrientation);
     
     //Setting the questionOrientaiton to zero to force vertical
     questionOrientation = 0;
